fix on amount value as 0 bug, fetching data from api but it breaks screen

+++ lib/components/operation_container.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
 
 class OperationContainer extends StatelessWidget {
-  final double amount;
+  final int amount;
   final String description;
   final bool isEntry;
   final String date;
